+++ dd-smoke-tests/src/main/groovy/datadog/smoketest/AbstractIastSpringBootSmokeTest.groovy
@@ -302,6 +302,18 @@ abstract class AbstractIastSpringBootSmokeTest extends AbstractServerSmokeTest {
     foundTaintedString
   }
 
+  def "weak hash vulnerability is present on boot"() {
+    setup:
+    String url = "http://localhost:${httpPort}/greeting"
+    def request = new Request.Builder().url(url).get().build()
+
+    when: 'ensure the controller is loaded'
+    client.newCall(request).execute()
+
+    then: 'a vulnerability pops in the logs (startup traces might not always be available)'
+    hasVulnerabilityInLogs(type('WEAK_HASH').and(evidence('SHA1')).and(withSpan()))
+  }
+
   private static Function<DecodedSpan, Boolean> hasMetric(final String name, final Object value) {
     return { span -> value == span.metrics.get(name) }
   }
@@ -364,5 +376,17 @@ abstract class AbstractIastSpringBootSmokeTest extends AbstractServerSmokeTest {
 
   private static String withSystemProperty(final String config, final Object value) {
     return "-Ddd.${config}=${value}"
+  }
+
+  boolean hasVulnerabilityInLogs(final Predicate<?> predicate) {
+    def found = false
+    checkLog { final String log ->
+      final index = log.indexOf(TAG_NAME)
+      if (index >= 0) {
+        final vulnerabilities = parseVulnerabilities(log, index)
+        found |= vulnerabilities.stream().anyMatch(predicate)
+      }
+    }
+    return found
   }
 }
